+++ CHANGES.rst
@@ -10,6 +10,8 @@ CHANGES
 - `FieldTracker` now respects `update_fields` changed in overridden `save()`
   method
 - Replace ugettext_lazy with gettext_lazy to satisfy Django deprecation warning
+- Add available_objects manager to SoftDeletableModel and add deprecation
+  warning to objects manager.
 
 4.0.0 (2019-12-11)
 ------------------